ledger: refactor ledger internal tests (#3149)
Summary
Refactor internal ledger evaluator testing to use the real ledger instead of using a dummy ledger.
In order to do this, I moved the related tests into an "internal_test" package. That allowed me to import both "ledger/internal" as well as "ledger", and conduct the test against a real ledger.

Test Plan
These are tests.
